+++ electron-vite-project/src/components/AddScene.tsx
@@ -53,9 +53,13 @@ function AddScene({ onClose }: AddSceneProps) {
     } else {
       // If the name is empty, the text box will be highlighted in red
       const textBox = document.getElementsByClassName('AddSceneTextBox')[0] as HTMLInputElement;
-      textBox.focus();
-      textBox.style.outline = '2px solid red';
-      textBox.style.outlineOffset = '-1px';
+      if (textBox) {
+        textBox.classList.add('error-outline');
+        textBox.focus();
+        setTimeout(() => {
+          textBox.classList.remove('error-outline');
+        }, 4000);
+      }
     }
   };